+++ CHANGES.md
@@ -6,6 +6,7 @@
 * ActiveModel validation: I18n support with country specific messages
 * ActiveModel validation: I18n locales in english and german
 * Fixed bug with wrong iso country code on greek vat number (EL != GR)
+* Valvat::Util.split only returns countries from europe
 
 ### 0.0.3 / 2011-01-06